pyproject.toml
[project]
name = "sinhala-ed-assistant"
version = "0.1.0"
description = "Sinhala Essay Evaluation System with AI Feedback"
readme = "README.md"
requires-python = ">=3.9"

dependencies = [
    "fastapi>=0.123.0",
    "uvicorn[standard]>=0.38.0",
    "sqlalchemy>=2.0.0",
    "psycopg2-binary>=2.9.9",
    "pgvector>=0.4.2",
    "sentence-transformers>=2.2.0",
    "rank-bm25>=0.2.0",
    "google-genai>=0.3.0",
    "google-api-python-client>=2.187.0",
    # Pin CPU version explicitly
    "torch>=2.4.0,<2.6.0",
    "torchvision>=0.19.0,<0.21.0",
    "numpy>=1.26.0,<2.0.0",  # Avoid numpy 2.x compatibility issues
    "pydantic>=2.12.0",
    "pydantic[email]>=2.12.0",
    "pydantic-settings>=2.0.0",
    "python-multipart>=0.0.6",
    "python-dotenv>=1.0.0",
    "firebase-admin>=6.0.0",
    "pillow>=12.0.0",
    "aiofiles>=23.0.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "opencv-python-headless>=4.10.0",  # Use headless version
    "scikit-learn>=1.7.0",
    "scipy>=1.15.0,<1.16.0; python_version < '3.11'",
    "scipy>=1.16.0; python_version >= '3.11'",
    "transformers>=4.57.0,<5.0.0",
    "tqdm>=4.67.0",
    "librosa>=0.11.0",
    "jiwer>=3.0.0",
    "soundfile>=0.13.0",
    "pdf2image>=1.17.0",
    "pdfplumber>=0.11.8",
    "PyMuPDF>=1.26.6",
    "pytesseract>=0.3.13",
    "ultralytics>=8.3.0",
    "alembic>=1.16.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.black]
line-length = 88
target-version = ['py312']  # Updated to Python 3.12
